docs: Added decorators to secure views

+++ products/views.py
@@ -3,6 +3,7 @@
 # 3rd party:
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -78,10 +79,15 @@ def product_details(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     A view for adding new product types and templates. Only administrative users can make changes.
     """
+    if not request.user.is_superuser:
+        messages.error(request, "This action can only be performed by super users.")
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -101,11 +107,16 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     A view for editing existing product types and templates.
     Only administrative users can make changes.
     """
+    if not request.user.is_superuser:
+        messages.error(request, "This action can only be performed by super users.")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -129,10 +140,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     A view to delete a product and template.
     """
+    if not request.user.is_superuser:
+        messages.error(request, "This action can only be performed by super users.")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product removed')
